assignments/pyproject.toml
[project]
name = "assignments"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.2.8",
    "requests>=2.32.5",
    "tavily-python>=0.7.11",
]
